backend/requirements.txt
# Django and Django REST Framework
Django==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
django-filter==25.1

# API Documentation
drf-spectacular==0.28.0

# Database & Caching
psycopg[binary]==3.2.9
redis==5.0.1
django-redis==5.4.0

# Task Queue
celery==5.3.6

# AI/ML
sentence-transformers==5.1.0
scikit-learn==1.6.1
numpy==2.2.3

# Image Processing
Pillow==11.0.0
reportlab==4.2.5
svglib==1.5.1

# Python
python-dotenv==1.0.1
requests==2.32.3
chardet==5.2.0

# Payments
stripe==11.5.0
